Implemented a Continuous Memory Allocator
This new allocator creates a sequence of anonymous maps that are
continuous not only in virtual memory, but also in physical memory.
For this, it requires access to /proc/self/pagemap, meaning that on
newer kenrels it needs to be run as root. Theoretically it is possible
wihtout admin capabilities, but it's very unlikely that it will work.

To implement the new allocator, I translated the codebase to C++,
renaming the victims and modifying the makefile in the process.

Integration with the eviction main thread will be done in a future
commit.

Signed-off-by: Radu Mantu <andru.mantu@gmail.com>